restructure lambdas
