+++ ops/aws/canary/lambda/pkg/models/types.go
@@ -1,0 +1,5 @@
+package models
+
+type Event struct {
+	Action string `json:"action"`
+}
