+++ ops/aws/canary/lambda/pkg/router/interfaces.go
@@ -1,0 +1,7 @@
+package router
+
+import (
+	"context"
+)
+
+type Handler func(ctx context.Context) error
