+++ ops/aws/canary/lambda/pkg/scenarios/submitAndDescribe.go
@@ -3,14 +3,18 @@ package scenarios
 import (
 	"context"
 	"fmt"
-	"github.com/filecoin-project/bacalhau/pkg/publicapi"
+	"github.com/filecoin-project/bacalhau/cmd/bacalhau"
 	"github.com/rs/zerolog/log"
 )
 
 // This scenario mainly calls the same client APIs that describe cmd does.
 // It doesn't do much validation except making sure the calls don't fail.
 // TODO: we should introduce a describe API and move all the logic there instead of cmd/describe.go
-func SubmitAnDescribe(ctx context.Context, client *publicapi.APIClient) error {
+func SubmitAnDescribe(ctx context.Context) error {
+	// intentionally delay creation of the client so a new client is created for each
+	// scenario to mimic the behavior of bacalhau cli.
+	client := bacalhau.GetAPIClient()
+
 	jobSpec, jobDeal := getSampleDockerJob()
 	submittedJob, err := client.Submit(ctx, jobSpec, jobDeal, nil)
 	log.Info().Msgf("submitted job: %s", submittedJob.ID)
