+++ ops/aws/canary/lambda/pkg/scenarios/types.go
@@ -1,0 +1,11 @@
+package scenarios
+
+import (
+	"context"
+)
+
+type Event struct {
+	Action string `json:"action"`
+}
+
+type Handler func(ctx context.Context) error
