+++ ops/aws/canary/lambda/pkg/scenarios/utils.go
